refactor: util
